Moved each UFCP format to a separate file.
This will keep it more organized in the future, since the ufcp.lua would increase its size significantly.
