+++ Cockpit/Scripts/Systems/ufcp_cf.lua
@@ -1,0 +1,19 @@
+dofile(LockOn_Options.script_path.."command_defs.lua")
+dofile(LockOn_Options.script_path.."functions.lua")
+dofile(LockOn_Options.script_path.."devices.lua")
+dofile(LockOn_Options.script_path.."Systems/avionics_api.lua")
+dofile(LockOn_Options.script_path.."utils.lua")
+dofile(LockOn_Options.script_path.."dump.lua")
+dofile(LockOn_Options.script_path.."Systems/ufcp_api.lua")
+
+-- Methods
+
+function update_c_f()
+    local text = ""
+    text = text .. "C/F\n"
+    UFCP_TEXT:set(text)
+end
+
+function SetCommandCF(command,value)
+
+end
