+++ Cockpit/Scripts/Systems/ufcp_navaids.lua
@@ -1,0 +1,18 @@
+dofile(LockOn_Options.script_path.."command_defs.lua")
+dofile(LockOn_Options.script_path.."functions.lua")
+dofile(LockOn_Options.script_path.."devices.lua")
+dofile(LockOn_Options.script_path.."Systems/avionics_api.lua")
+dofile(LockOn_Options.script_path.."utils.lua")
+dofile(LockOn_Options.script_path.."dump.lua")
+dofile(LockOn_Options.script_path.."Systems/ufcp_api.lua")
+
+-- NAV AIDS
+function update_nav_aids()
+    local text = ""
+    text = text .. "NAV AIDS\n"
+    UFCP_TEXT:set(text)
+end
+
+function SetCommandNavAids(command,value)
+
+end
