+++ Cockpit/Scripts/Systems/ufcp_oap.lua
@@ -1,0 +1,19 @@
+dofile(LockOn_Options.script_path.."command_defs.lua")
+dofile(LockOn_Options.script_path.."functions.lua")
+dofile(LockOn_Options.script_path.."devices.lua")
+dofile(LockOn_Options.script_path.."Systems/avionics_api.lua")
+dofile(LockOn_Options.script_path.."utils.lua")
+dofile(LockOn_Options.script_path.."dump.lua")
+dofile(LockOn_Options.script_path.."Systems/ufcp_api.lua")
+
+--methods
+
+function update_oap()
+    local text = ""
+    text = text .. "OAP\n"
+    UFCP_TEXT:set(text)
+end
+
+function SetCommandOAP(command,value)
+
+end
