+++ Cockpit/Scripts/Systems/ufcp_tac.lua
@@ -1,0 +1,48 @@
+dofile(LockOn_Options.script_path.."command_defs.lua")
+dofile(LockOn_Options.script_path.."functions.lua")
+dofile(LockOn_Options.script_path.."devices.lua")
+dofile(LockOn_Options.script_path.."Systems/avionics_api.lua")
+dofile(LockOn_Options.script_path.."utils.lua")
+dofile(LockOn_Options.script_path.."dump.lua")
+dofile(LockOn_Options.script_path.."Systems/ufcp_api.lua")
+
+-- Methods
+
+function update_tac()
+    local text = ""
+    text = text .. "TAC MENU\n"
+    text = text .. "1CTLN 2AVOID 3        \n"
+    text = text .. "4     5      6        \n"
+    text = text .. "7     8      9        \n"
+    UFCP_TEXT:set(text)
+end
+
+function SetCommandTacMenu(command,value)
+    if command == device_commands.UFCP_1 and value == 1 then
+        ufcp_sel_format = UFCP_FORMAT_IDS.TAC_CTLN
+    elseif command == device_commands.UFCP_2 and value == 1 then
+        ufcp_sel_format = UFCP_FORMAT_IDS.TAC_AVOID
+    end
+end
+
+-- CTLN
+function update_tac_ctln()
+    local text = ""
+    text = text .. "TAC CTLN\n"
+    UFCP_TEXT:set(text)
+end
+
+function SetCommandTacCtln(command,value)
+
+end
+
+-- AVOID
+function update_tac_avoid()
+    local text = ""
+    text = text .. "TAC AVOID\n"
+    UFCP_TEXT:set(text)
+end
+
+function SetCommandTacAvoid(command,value)
+
+end
